Move TimetablesDataset.getNested to an extension method
So it can be called for any mutableMap, not just the dataset.
This allows more flexible usage in tests.
